feat/fix: minibatch gd implementation
Found that minibatch gd implementation actually performs much better and more accurate than minibatch adam implementation.

This is probably due to the fact that the dataset is not variant/random enough to make Adam work effectively. As a result, Adam implementation makes the loss increase back up with time and creates sudden oscillations.